+++ client/CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [v0.5.0 Unreleased]
 
 ### Added
+- Added use-debounce library 
 - Integration of local layer to handle different type of layers in the future. [LANDGRIF-1248](https://vizzuality.atlassian.net/browse/LANDGRIF-1248)
 - New purple style for scenario select in analysis page. [LANDGRIF-1037](https://vizzuality.atlassian.net/browse/LANDGRIF-1037)
 - Add filter by year on analysis table. [LANDGRIF-965](https://vizzuality.atlassian.net/browse/LANDGRIF-965)
+### Changed
+- Changed map component to include view state change debounce
+- Changed impact layer resolution to accomodate zoom value [LANDGRIF-1131](https://vizzuality.atlassian.net/browse/LANDGRIF-1131)
 
 ### Changed
 - Updated axios version.